feat: update Badges component text size and add SocialLinks component with animated icons
